[Peripherals] Move events to subfolder
