+++ xbmc/games/agents/input/AgentInput.cpp
@@ -14,10 +14,10 @@
 #include "games/addons/input/GameClientJoystick.h"
 #include "games/controllers/Controller.h"
 #include "input/InputManager.h"
-#include "peripherals/EventLockHandle.h"
 #include "peripherals/Peripherals.h"
 #include "peripherals/devices/Peripheral.h"
 #include "peripherals/devices/PeripheralJoystick.h"
+#include "peripherals/events/EventLockHandle.h"
 #include "utils/log.h"
 
 #include <array>
