+++ xbmc/peripherals/CMakeLists.txt
@@ -1,13 +1,6 @@
-set(SOURCES EventLockHandle.cpp
-            EventPollHandle.cpp
-            EventScanner.cpp
-            Peripherals.cpp)
+set(SOURCES Peripherals.cpp)
 
-set(HEADERS EventLockHandle.h
-            EventPollHandle.h
-            EventScanner.h
-            IEventScannerCallback.h
-            Peripherals.h
+set(HEADERS Peripherals.h
             PeripheralTypes.h)
 
 core_add_library(peripherals)
