+++ xbmc/peripherals/events/EventLockHandle.h
