+++ xbmc/peripherals/events/EventPollHandle.h
