+++ xbmc/peripherals/events/EventScanner.h
@@ -10,7 +10,7 @@
 
 #include "EventLockHandle.h"
 #include "EventPollHandle.h"
-#include "PeripheralTypes.h"
+#include "peripherals/PeripheralTypes.h"
 #include "threads/CriticalSection.h"
 #include "threads/Event.h"
 #include "threads/Thread.h"
