+++ xbmc/peripherals/events/interfaces/CMakeLists.txt
@@ -1,0 +1,6 @@
+set(HEADERS IEventScannerCallback.h
+)
+
+if(NOT ENABLE_STATIC_LIBS)
+  core_add_library(peripherals_events_interfaces)
+endif()
