+++ xbmc/peripherals/events/interfaces/IEventScannerCallback.h
